fix(router): pop bugs in the payments confirm flow
If the customer explicitly triggers a payments confirm flow, we want to either get a payment method in the request or a payment token which maps to the chosen payment method. Also we want to ensure that we HAVE the payment method before we ever get to the update_trackers step.