ci: temporarily disable path-based filtering
Currently, if a PR has changes to both Rust code and other files which are not tested on CI, GitHub Actions seems to just skip tests for the whole PR, which is not intended. Thus, the path-based filtering is being disabled temporarily until a more elegant solution for the problem has been found.