fix(cypress): fix cypress throwing cypressError w.r.t cy.task()
The error that we get when running Cypress test through UI is as below:
Error:    CypressError: cy.task() must only be invoked from the spec file or support file.

The reason being cy.task neither being in helper file nor a spec file.
Moved it back to where it belong to.