feat: rename tests and add example payment intent tests
0001 contains an example for payment intent create call, confirm call
and retrieve call. and of course, these does not work at present and
hence, skipped intentionally from being run.

add more number contexts within describe to address all the cases
mentioned at the top of the file.

configs needs to be introduced. but before that, we need to do some
structural changes to change the behavior