refactor(cypress): do not update `card_expiry` while updating card info
the reason being there are a few connectors like adyen that expects the
user to pass specific card expiry year and having it hard coded results
in unintended issues.

we cannot hard code connector names in spec files. neither we can go
ahead and pull in another set of connector configs for this.

so, it is more viable to remove the entire field from getting update
since updating one field should be more than enough.